@@ -31,58 +31,96 @@
     <properties>
         <start-class>manon.Application</start-class>
 
-        <java.version>1.8</java.version>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
+        <java.version>1.9</java.version>
+        <maven.compiler.source>1.9</maven.compiler.source>
+        <maven.compiler.target>1.9</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 
         <!-- Annotation -->
-        <lombok.version>1.16.18</lombok.version>
+        <lombok.version>1.16.20</lombok.version>
+        <jetbrains.annotations.version>15.0</jetbrains.annotations.version>
 
         <!-- Misc -->
         <streamex.version>0.6.5</streamex.version>
         <jool.version>0.9.12</jool.version>
+        <kotlin-stdlib.version>1.2.10</kotlin-stdlib.version>
 
         <!-- Testing -->
-        <assertj-core.version>3.8.0</assertj-core.version>
+        <assertj-core.version>3.9.0</assertj-core.version>
         <hamcrest-library.version>1.3</hamcrest-library.version>
-        <janino.version>3.0.7</janino.version>
-        <testng.version>6.11</testng.version>
-        <rest-assured.version>2.9.0</rest-assured.version>
-        <mockito-core.version>2.11.0</mockito-core.version>
+        <janino.version>3.0.8</janino.version>
+        <testng.version>6.13.1</testng.version>
+        <rest-assured.version>3.0.6</rest-assured.version>
+        <mockito-core.version>2.13.0</mockito-core.version>
 
         <!-- JSON -->
-        <jackson.version>2.9.2</jackson.version>
-        <jackson-databind.version>2.9.2</jackson-databind.version>
-
-        <!-- MongoDB -->
-        <mongo.version>3.4.2</mongo.version>
+        <jackson.version>2.9.3</jackson.version>
+        <jackson-annotations.version>2.9.3</jackson-annotations.version>
+        <jackson-databind.version>2.9.3</jackson-databind.version>
+        <jackson-module-kotlin.version>2.9.3</jackson-module-kotlin.version>
 
         <!-- SQL -->
-        <mariadb.version>2.1.2</mariadb.version>
+        <mariadb.version>2.2.1</mariadb.version>
         <hsqldb.version>2.4.0</hsqldb.version>
 
         <mysema-querydsl.version>3.7.4</mysema-querydsl.version>
 
         <!-- Log -->
         <jboss-logging.version>3.3.1.Final</jboss-logging.version>
-        <log4j.version>1.2.17</log4j.version>
         <logback-classic.version>1.2.3</logback-classic.version>
         <slf4j-api.version>1.7.25</slf4j-api.version>
 
         <!-- Aspect -->
         <jcabi-aspects.version>0.22.5</jcabi-aspects.version>
-        <aspectjrt.version>1.8.12</aspectjrt.version>
+        <aspectjrt.version>1.9.0.RC2</aspectjrt.version>
 
         <!-- Spring -->
-        <spring-boot.version>1.5.8.RELEASE</spring-boot.version>
-        <spring-framework.version>4.3.12.RELEASE</spring-framework.version>
-        <spring-security-config.version>4.2.3.RELEASE</spring-security-config.version>
-        <spring-security-web.version>4.2.3.RELEASE</spring-security-web.version>
-        <spring-batch-test.version>3.0.8.RELEASE</spring-batch-test.version>
+        <spring-boot.version>2.0.0.M7</spring-boot.version>
+        <spring-framework.version>5.0.2.RELEASE</spring-framework.version>
+        <spring-security-config.version>5.0.0.RELEASE</spring-security-config.version>
+        <spring-security-web.version>5.0.0.RELEASE</spring-security-web.version>
+        <spring-batch-test.version>4.0.0.RELEASE</spring-batch-test.version>
 
     </properties>
+
+    <repositories>
+        <repository>
+            <id>jcenter</id>
+            <name>jcenter</name>
+            <url>http://jcenter.bintray.com/</url>
+            <snapshots>
+                <enabled>false</enabled>
+            </snapshots>
+        </repository>
+        <repository>
+            <id>spring-milestones</id>
+            <name>Spring Milestones</name>
+            <url>https://repo.spring.io/milestone</url>
+            <snapshots>
+                <enabled>false</enabled>
+            </snapshots>
+        </repository>
+    </repositories>
+
+    <pluginRepositories>
+        <pluginRepository>
+            <id>jcenter</id>
+            <name>jcenter</name>
+            <url>http://jcenter.bintray.com/</url>
+            <snapshots>
+                <enabled>false</enabled>
+            </snapshots>
+        </pluginRepository>
+        <pluginRepository>
+            <id>spring-milestones</id>
+            <name>Spring Milestones</name>
+            <url>https://repo.spring.io/milestone</url>
+            <snapshots>
+                <enabled>false</enabled>
+            </snapshots>
+        </pluginRepository>
+    </pluginRepositories>
 
     <dependencies>
 
