@@ -4,7 +4,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.converter.HttpMessageConverter;
 import org.springframework.http.converter.json.AbstractJackson2HttpMessageConverter;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 import java.text.SimpleDateFormat;
 import java.util.List;
@@ -13,7 +14,8 @@ import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 import static manon.util.Tools.DATE_FORMAT;
 
 @Configuration
-public class ConverterConfig extends WebMvcConfigurerAdapter {
+@EnableWebMvc
+public class ConverterConfig implements WebMvcConfigurer {
     
     @Override
     public void extendMessageConverters(List<HttpMessageConverter<?>> converters) {
@@ -23,6 +25,5 @@ public class ConverterConfig extends WebMvcConfigurerAdapter {
             objectMapper.setDateFormat(new SimpleDateFormat(DATE_FORMAT));
             objectMapper.setSerializationInclusion(NON_NULL);
         });
-        super.extendMessageConverters(converters);
     }
 }
