@@ -1,6 +1,7 @@
 package manon.matchmaking.repository;
 
 import manon.matchmaking.document.LobbyTeam;
+import org.jetbrains.annotations.NotNull;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
@@ -9,7 +10,7 @@ import java.util.Optional;
 @Repository
 public interface LobbyTeamRepository extends MongoRepository<LobbyTeam, String>, LobbyTeamRepositoryCustom {
     
-    Optional<LobbyTeam> findById(String id);
+    @NotNull Optional<LobbyTeam> findById(@NotNull String id);
     
-    Optional<LobbyTeam> findByProfileIds(String profileId);
+    Optional<LobbyTeam> findByUserIds(String userId);
 }
