@@ -1,13 +1,11 @@
 package manon.user;
 
+import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+@AllArgsConstructor
 @Getter
 public class UserExistsException extends Exception {
     
     private String username;
-    
-    public UserExistsException(String username) {
-        this.username = username;
-    }
 }
