@@ -1,9 +1,13 @@
 package manon.user.service;
 
+import com.mongodb.DuplicateKeyException;
 import manon.user.UserExistsException;
 import manon.user.UserNotFoundException;
 import manon.user.document.User;
+import manon.user.form.UserFieldEnum;
+import manon.user.form.UserUpdateForm;
 import manon.user.registration.RegistrationStateEnum;
+import manon.user.repository.UserRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -12,47 +16,36 @@ import java.util.Optional;
 
 public interface UserService {
     
-    /**
-     * Get all users.
-     * @param pageable page request.
-     * @return paginated list of users.
-     */
-    Page<User> findAll(Pageable pageable);
+    long count();
     
-    /**
-     * Get a user by username.
-     * @param username user username.
-     * @return user.
-     */
+    void save(User user);
+    
+    void ensureExist(String... ids) throws UserNotFoundException;
+    
+    User readOne(String id) throws UserNotFoundException;
+    
+    User readByUsername(String username) throws UsernameNotFoundException;
+    
     Optional<User> findByUsername(String username);
     
     /**
-     * Get a user by username.
-     * @param username user username.
-     * @return user.
+     * Update a user's field.
+     * @see UserRepository#updateField(String, UserFieldEnum, Object)
+     * @param userId user id.
+     * @param userUpdateForm field name and its new value.
      */
-    User readByUsername(String username)
-            throws UsernameNotFoundException;
+    void update(String userId, UserUpdateForm userUpdateForm) throws UserNotFoundException, DuplicateKeyException;
+    
+    Page<User> findAll(Pageable pageable);
     
     /**
-     * Get a user by id.
-     * @param id user id.
-     * @return user.
-     */
-    User readOne(String id)
-            throws UserNotFoundException;
-    
-    /**
-     * Create a new user and its profile.
+     * Create a new user and its user.
      * @param user user data.
      * @return new user.
      */
-    User create(User user)
-            throws UserExistsException;
+    User create(User user) throws UserExistsException;
     
-    void setPassword(String id, String password)
-            throws UserNotFoundException;
+    void setPassword(String id, String password) throws UserNotFoundException;
     
-    void setRegistrationState(String id, RegistrationStateEnum registrationState)
-            throws UserNotFoundException;
+    void setRegistrationState(String id, RegistrationStateEnum registrationState) throws UserNotFoundException;
 }
